refactor: 테스트 클래스 패키지 분리
