+++ src/test/java/nextstep/auth/application/EmailPasswordAuthenticationServiceTest.java
@@ -1,4 +1,4 @@
-package nextstep.auth.unit;
+package nextstep.auth.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
