+++ src/test/java/nextstep/subway/line/application/LineServiceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.unit;
+package nextstep.subway.line.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
