+++ src/test/java/nextstep/subway/line/domain/entity/LineSectionTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.unit;
+package nextstep.subway.line.domain.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
