+++ src/test/java/nextstep/subway/path/web/PathControllerTest.java
@@ -1,0 +1,7 @@
+package nextstep.subway.path.web;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class PathControllerTest {
+
+}
